+++ Ecosens_WebPage/Models/DashboardViewModel.cs
@@ -5,10 +5,10 @@ namespace Ecosens_WebPage.Models
     public class DashboardViewModel
     {
         [JsonPropertyName("plastico")]
-        public int Plastico { get; set; }
+        public double Plastico { get; set; }
         [JsonPropertyName("metal")]
-        public int Metal { get; set; }
-        public int Total => Plastico + Metal;
+        public double Metal { get; set; }
+        public double Total => Plastico + Metal;
         public int TotalAlertas { get; set; }
     }
 }
